refactor: transactions
