+++ backend/lcfs/web/api/transaction/views.py
@@ -1,20 +1,16 @@
 import math
-import io
-from datetime import datetime
+from logging import getLogger
 from fastapi import APIRouter, Depends, HTTPException, status, Body
 from fastapi.responses import StreamingResponse
-from sqlalchemy import func, select, distinct
-from sqlalchemy.orm import joinedload
 from sqlalchemy.ext.asyncio import AsyncSession
 from starlette.responses import Response
 from lcfs.db import dependencies
-from lcfs.db.models.Transaction import Transaction
-from lcfs.db.models.IssuanceHistory import IssuanceHistory
-from lcfs.db.models.TransferHistory import TransferHistory
 from lcfs.web.api.base import PaginationRequestSchema, PaginationResponseSchema
 from lcfs.web.api.transaction.schema import Transactions
 from lcfs.web.core.decorators import roles_required
-from lcfs.utils.spreadsheet_builder import SpreadsheetBuilder
+from lcfs.web.api.transaction.session import TransactionRepo
+
+logger = getLogger("transaction")
 
 router = APIRouter()
 get_async_db = dependencies.get_async_db_session
@@ -22,40 +18,20 @@ get_async_db = dependencies.get_async_db_session
 
 @router.get("/", status_code=status.HTTP_200_OK, response_model=Transactions)
 async def get_all_transactions(
-    db: AsyncSession = Depends(get_async_db),
     pagination: PaginationRequestSchema = Body(..., embed=False),
     response: Response = None,
+    repo: TransactionRepo = Depends(),
 ):
+    """
+    Endpoint to get all transactions along with the 'from' and 'to' organization data.
+    This will return paginated data.
+    """
     try:
-        offset = 0 if (pagination.page < 1) else (pagination.page - 1) * pagination.size
-        limit = pagination.size
 
-        query = select(Transaction).options(
-            joinedload(Transaction.issuance_history_record).options(
-                joinedload(IssuanceHistory.organization),
-                joinedload(IssuanceHistory.issuance_status),
-            ),
-            joinedload(Transaction.transfer_history_record).options(
-                joinedload(TransferHistory.to_organization),
-                joinedload(TransferHistory.from_organization),
-                joinedload(TransferHistory.transfer_status),
-            ),
-            joinedload(Transaction.transaction_type),
-        )
-        count_query = await db.execute(
-            select(func.count(distinct(Transaction.transaction_id)))
-        )
+        transactions, total_count = await repo.get_transactions(pagination)
 
-        total_count = count_query.unique().scalar_one_or_none()
-
-        transaction_results = await db.execute(query.offset(offset).limit(limit))
-        results = transaction_results.scalars().unique().all()
-
-        transactions = [
-            Transaction.model_validate(transaction) for transaction in results
-        ]
-
-        if len(transactions) == 0:
+        if not transactions:
+            logger.error("Error getting transactions")
             response.status_code = status.HTTP_404_NOT_FOUND
             return Transactions(
                 pagination=PaginationResponseSchema(
@@ -74,6 +50,7 @@ async def get_all_transactions(
         )
 
     except Exception as e:
+        logger.error(f"Error getting transactions: {str(e)}")
         response.status_code = status.HTTP_500_INTERNAL_SERVER_ERROR
         raise HTTPException(
             status_code=500,
@@ -83,54 +60,28 @@ async def get_all_transactions(
 
 @router.get("/export", response_class=StreamingResponse, status_code=status.HTTP_200_OK)
 @roles_required("Government")
-async def export_organizations(db: AsyncSession = Depends(get_async_db)):
-    export_format = "xls"
-    media_type = "application/vnd.ms-excel"
+async def export_organizations(db: AsyncSession = Depends(get_async_db), repo: TransactionRepo = Depends(),):
+    """
+    Endpoint to export information of transactions
+
+    This endpoint can support exporting data in different file formats (xls, xlsx, csv)
+    as specified by the 'export_format' and 'media_type' variables.
+    - 'export_format' specifies the file format: options are 'xls', 'xlsx', and 'csv'.
+    - 'media_type' sets the appropriate MIME type based on 'export_format':
+        'application/vnd.ms-excel' for 'xls',
+        'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet' for 'xlsx',
+        'text/csv' for 'csv'.
+
+    The SpreadsheetBuilder class is used for building the spreadsheet.
+    It allows adding multiple sheets with custom styling options and exports them as a byte stream.
+    Also, an example of how to use the SpreadsheetBuilder is provided in its class documentation.
+
+    Note: Only the first sheet data is used for the CSV format,
+        as CSV files do not support multiple sheets.
+    """
 
     try:
-        # Fetch all organizations from the database
-        result = await db.execute(
-            select(Transaction)
-            .options(joinedload(Transaction.transaction_type))
-            .order_by(Transaction.transaction_id)
-        )
-        transactions = result.scalars().all()
-
-        # Prepare data for the spreadsheet
-        data = [
-            [
-                transaction.transaction_id,
-                123,
-                transaction.transaction_type.status.value,
-            ]
-            for transaction in transactions
-        ]
-
-        # Create a spreadsheet
-        builder = SpreadsheetBuilder(file_format=export_format)
-
-        builder.add_sheet(
-            sheet_name="Transactions",
-            columns=[
-                "ID",
-                "Compliance Units",
-                "Status",
-            ],
-            rows=data,
-            styles={"bold_headers": True},
-        )
-
-        file_content = builder.build_spreadsheet()
-
-        # Get the current date in YYYY-MM-DD format
-        current_date = datetime.now().strftime("%Y-%m-%d")
-
-        filename = f"BC-LCFS-transactions-{current_date}.{export_format}"
-        headers = {"Content-Disposition": f'attachment; filename="{filename}"'}
-
-        return StreamingResponse(
-            io.BytesIO(file_content), media_type=media_type, headers=headers
-        )
+        return await repo.export_transactions()
 
     except Exception as e:
         raise HTTPException(
