+++ frontend/src/views/Transactions/data.js
@@ -1,0 +1,295 @@
+export const dummy = {
+  // only populated with the fields we need. actually query will probably return the full data.
+  transactions: [
+    {
+      transaction_id: 1,
+      compliance_period: '2024',
+      transaction_type: {
+        type: 'Transfer'
+      },
+      compliance_units: 123,
+      value_per_unit: 123,
+      last_updated: '2024-01-04T20:01:40.760Z',
+      issuance_history_record: {
+        organization: {
+          name: 'QuantumNova Fuels'
+        },
+        issuance_status: {
+          status: 'Recorded'
+        }
+      },
+      transfer_history_record: {
+        transfer_status: {
+          status: 'Recorded'
+        },
+        from_organization: {
+          name: 'NebulaWings Dynamics'
+        },
+        to_organization: {
+          name: 'QuantumNova Fuels'
+        }
+      }
+    },
+    {
+      transaction_id: 2,
+      compliance_period: '2024',
+      transaction_type: {
+        type: 'Transfer'
+      },
+      compliance_units: 123,
+      value_per_unit: 123,
+      last_updated: '2024-01-04T20:01:40.760Z',
+      issuance_history_record: {
+        organization: {
+          name: 'QuantumNova Fuels'
+        },
+        issuance_status: {
+          status: 'Recorded'
+        }
+      },
+      transfer_history_record: {
+        transfer_status: {
+          status: 'Recorded'
+        },
+        from_organization: {
+          name: 'NebulaWings Dynamics'
+        },
+        to_organization: {
+          name: 'QuantumNova Fuels'
+        }
+      }
+    },
+    {
+      transaction_id: 3,
+      compliance_period: '2024',
+      transaction_type: {
+        type: 'Transfer'
+      },
+      compliance_units: 123,
+      value_per_unit: 123,
+      last_updated: '2024-01-04T20:01:40.760Z',
+      issuance_history_record: {
+        organization: {
+          name: 'QuantumNova Fuels'
+        },
+        issuance_status: {
+          status: 'Recorded'
+        }
+      },
+      transfer_history_record: {
+        transfer_status: {
+          status: 'Recorded'
+        },
+        from_organization: {
+          name: 'NebulaWings Dynamics'
+        },
+        to_organization: {
+          name: 'QuantumNova Fuels'
+        }
+      }
+    },
+    {
+      transaction_id: 4,
+      compliance_period: '2024',
+      transaction_type: {
+        type: 'Transfer'
+      },
+      compliance_units: 123,
+      value_per_unit: 123,
+      last_updated: '2024-01-04T20:01:40.760Z',
+      issuance_history_record: {
+        organization: {
+          name: 'QuantumNova Fuels'
+        },
+        issuance_status: {
+          status: 'Recorded'
+        }
+      },
+      transfer_history_record: {
+        transfer_status: {
+          status: 'Recorded'
+        },
+        from_organization: {
+          name: 'NebulaWings Dynamics'
+        },
+        to_organization: {
+          name: 'QuantumNova Fuels'
+        }
+      }
+    },
+    {
+      transaction_id: 5,
+      compliance_period: '2024',
+      transaction_type: {
+        type: 'Transfer'
+      },
+      compliance_units: 123,
+      value_per_unit: 123,
+      last_updated: '2024-01-04T20:01:40.760Z',
+      issuance_history_record: {
+        organization: {
+          name: 'QuantumNova Fuels'
+        },
+        issuance_status: {
+          status: 'Recorded'
+        }
+      },
+      transfer_history_record: {
+        transfer_status: {
+          status: 'Recorded'
+        },
+        from_organization: {
+          name: 'NebulaWings Dynamics'
+        },
+        to_organization: {
+          name: 'QuantumNova Fuels'
+        }
+      }
+    },
+    {
+      transaction_id: 6,
+      compliance_period: '2024',
+      transaction_type: {
+        type: 'Transfer'
+      },
+      compliance_units: 123,
+      value_per_unit: 123,
+      last_updated: '2024-01-04T20:01:40.760Z',
+      issuance_history_record: {
+        organization: {
+          name: 'QuantumNova Fuels'
+        },
+        issuance_status: {
+          status: 'Recorded'
+        }
+      },
+      transfer_history_record: {
+        transfer_status: {
+          status: 'Recorded'
+        },
+        from_organization: {
+          name: 'NebulaWings Dynamics'
+        },
+        to_organization: {
+          name: 'QuantumNova Fuels'
+        }
+      }
+    },
+    {
+      transaction_id: 7,
+      compliance_period: '2024',
+      transaction_type: {
+        type: 'Transfer'
+      },
+      compliance_units: 123,
+      value_per_unit: 123,
+      last_updated: '2024-01-04T20:01:40.760Z',
+      issuance_history_record: {
+        organization: {
+          name: 'QuantumNova Fuels'
+        },
+        issuance_status: {
+          status: 'Recorded'
+        }
+      },
+      transfer_history_record: {
+        transfer_status: {
+          status: 'Recorded'
+        },
+        from_organization: {
+          name: 'NebulaWings Dynamics'
+        },
+        to_organization: {
+          name: 'QuantumNova Fuels'
+        }
+      }
+    },
+    {
+      transaction_id: 8,
+      compliance_period: '2024',
+      transaction_type: {
+        type: 'Transfer'
+      },
+      compliance_units: 123,
+      value_per_unit: 123,
+      last_updated: '2024-01-04T20:01:40.760Z',
+      issuance_history_record: {
+        organization: {
+          name: 'QuantumNova Fuels'
+        },
+        issuance_status: {
+          status: 'Recorded'
+        }
+      },
+      transfer_history_record: {
+        transfer_status: {
+          status: 'Recorded'
+        },
+        from_organization: {
+          name: 'NebulaWings Dynamics'
+        },
+        to_organization: {
+          name: 'QuantumNova Fuels'
+        }
+      }
+    },
+    {
+      transaction_id: 9,
+      compliance_period: '2024',
+      transaction_type: {
+        type: 'Transfer'
+      },
+      compliance_units: 123,
+      value_per_unit: 123,
+      last_updated: '2024-01-04T20:01:40.760Z',
+      issuance_history_record: {
+        organization: {
+          name: 'QuantumNova Fuels'
+        },
+        issuance_status: {
+          status: 'Recorded'
+        }
+      },
+      transfer_history_record: {
+        transfer_status: {
+          status: 'Recorded'
+        },
+        from_organization: {
+          name: 'NebulaWings Dynamics'
+        },
+        to_organization: {
+          name: 'QuantumNova Fuels'
+        }
+      }
+    },
+    {
+      transaction_id: 10,
+      compliance_period: '2024',
+      transaction_type: {
+        type: 'Transfer'
+      },
+      compliance_units: 123,
+      value_per_unit: 123,
+      last_updated: '2024-01-04T20:01:40.760Z',
+      issuance_history_record: {
+        organization: {
+          name: 'QuantumNova Fuels'
+        },
+        issuance_status: {
+          status: 'Recorded'
+        }
+      },
+      transfer_history_record: {
+        transfer_status: {
+          status: 'Recorded'
+        },
+        from_organization: {
+          name: 'NebulaWings Dynamics'
+        },
+        to_organization: {
+          name: 'QuantumNova Fuels'
+        }
+      }
+    }
+  ]
+}
