+++ frontend/src/views/Transactions/index.js
@@ -1,0 +1,1 @@
+export { Transactions } from './Transactions'
